integrated flexbox with the widgets leveraging reusable classes
