+++ src/js/components/App.tsx
@@ -22,6 +22,9 @@ const App = (): JSX.Element => {
 
   return (
     <>
+      {/* <div
+      className="dim-container"
+    ></div> */}
       <Header />
       <MapContent />
     </>
