--- src/js/components/mapControls/printWidget.tsx
@@ -1,52 +1,0 @@
-import React, { FunctionComponent, useState } from 'react';
-
-import ModalCard from '../modal/modalCard';
-
-// import '../../../css/shareWidget'
-
-const PrintWidget: FunctionComponent = () => {
-  const [openWidget, setOpenWidget] = useState(false);
-  // TODO integrate Redux
-
-  const WidgetContent = () => {
-    return (
-      <div className="share-content-container">
-        <div className="directions">
-          <p>Choose a print output</p>
-        </div>
-      </div>
-    );
-  };
-
-  return (
-    <>
-      <div className="print-widget-container left">
-        <button
-          className="image-wrapper"
-          role="button"
-          aria-pressed={openWidget}
-          onClick={() => setOpenWidget(!openWidget)}
-        >
-          <svg className="svg-icon">
-            <svg id="icon-print" viewBox="0 0 19 14">
-              <title>Print</title>
-              <path
-                d="M16.000,10.000 C16.000,10.000 16.000,12.000 16.000,12.000 C16.000,12.000 16.000,13.000 16.000,13.000 C16.000,13.000 16.000,14.000 16.000,14.000 C16.000,14.000 3.000,14.000 3.000,14.000 C3.000,14.000 3.000,13.000 3.000,13.000 C3.000,13.000 3.000,12.000 3.000,12.000 C3.000,12.000 3.000,10.000 3.000,10.000 C3.000,10.000 0.000,10.000 0.000,10.000 C0.000,10.000 0.000,3.000 0.000,3.000 C0.000,3.000 1.000,3.000 1.000,3.000 C1.000,3.000 2.000,3.000 2.000,3.000 C2.000,3.000 3.000,3.000 3.000,3.000 C3.000,3.000 3.000,0.000 3.000,0.000 C3.000,0.000 16.000,0.000 16.000,0.000 C16.000,0.000 16.000,3.000 16.000,3.000 C16.000,3.000 17.000,3.000 17.000,3.000 C17.000,3.000 18.000,3.000 18.000,3.000 C18.000,3.000 19.000,3.000 19.000,3.000 C19.000,3.000 19.000,10.000 19.000,10.000 C19.000,10.000 16.000,10.000 16.000,10.000 ZM5.000,12.000 C5.000,12.000 14.000,12.000 14.000,12.000 C14.000,12.000 14.000,7.993 14.000,7.993 C14.000,7.993 5.000,7.993 5.000,7.993 C5.000,7.993 5.000,12.000 5.000,12.000 ZM2.889,8.002 C2.889,8.002 2.889,8.000 2.889,8.000 C2.889,8.000 3.000,8.000 3.000,8.000 C3.000,8.000 3.000,7.000 3.000,7.000 C3.000,7.000 3.000,6.000 3.000,6.000 C3.000,6.000 16.000,6.000 16.000,6.000 C16.000,6.000 16.000,7.000 16.000,7.000 C16.000,7.000 16.000,8.000 16.000,8.000 C16.000,8.000 16.111,8.000 16.111,8.000 C16.111,8.000 16.111,8.002 16.111,8.002 C16.111,8.002 17.000,8.002 17.000,8.002 C17.000,8.002 17.000,5.000 17.000,5.000 C17.000,5.000 2.000,5.000 2.000,5.000 C2.000,5.000 2.000,8.002 2.000,8.002 C2.000,8.002 2.889,8.002 2.889,8.002 ZM1.000,3.996 C1.000,3.996 1.000,3.996 1.000,3.996 C1.000,3.996 1.000,8.002 1.000,8.002 C1.000,8.002 1.000,8.002 1.000,8.002 C1.000,8.002 1.000,5.000 1.000,5.000 C1.000,5.000 1.000,4.000 1.000,4.000 C1.000,4.000 1.000,3.996 1.000,3.996 ZM2.000,3.996 C2.000,3.996 2.000,4.000 2.000,4.000 C2.000,4.000 17.000,4.000 17.000,4.000 C17.000,4.000 17.000,3.996 17.000,3.996 C17.000,3.996 2.000,3.996 2.000,3.996 ZM18.000,3.996 C18.000,3.996 18.000,3.996 18.000,3.996 C18.000,3.996 18.000,4.000 18.000,4.000 C18.000,4.000 18.000,5.000 18.000,5.000 C18.000,5.000 18.000,8.002 18.000,8.002 C18.000,8.002 18.000,8.002 18.000,8.002 C18.000,8.002 18.000,3.996 18.000,3.996 Z"
-                fill-rule="evenodd"
-              ></path>
-            </svg>
-          </svg>
-        </button>
-      </div>
-
-      <ModalCard
-        renderModal={openWidget}
-        setOpenWidget={(modalStatus: boolean) => setOpenWidget(modalStatus)}
-      >
-        <WidgetContent />
-      </ModalCard>
-    </>
-  );
-};
-
-export default PrintWidget;
