+++ src/js/components/mapWidgets/printWidget.tsx
@@ -2,9 +2,9 @@ import React, { FunctionComponent, useState } from 'react';
 
 import ModalCard from '../modal/modalCard';
 
-import { ReactComponent as ShareIcon } from '../../../images/ShareIcon.svg';
+import { ReactComponent as PrintIcon } from '../../../images/PrintIcon.svg';
 
-const ShareWidget: FunctionComponent = () => {
+const PrintWidget: FunctionComponent = () => {
   const [openWidget, setOpenWidget] = useState(false);
   // TODO integrate Redux
 
@@ -12,11 +12,7 @@ const ShareWidget: FunctionComponent = () => {
     return (
       <div className="share-content-container">
         <div className="directions">
-          <h4 className="title">Share this view</h4>
-          <p>
-            Copy and paste the link to share it or use the buttons below to
-            share on social media.
-          </p>
+          <p>Choose a print output</p>
         </div>
       </div>
     );
@@ -24,16 +20,17 @@ const ShareWidget: FunctionComponent = () => {
 
   return (
     <>
-      <div className="share-widget-container right">
+      <div className="widget-container">
         <button
           className="image-wrapper"
           role="button"
           aria-pressed={openWidget}
           onClick={() => setOpenWidget(!openWidget)}
         >
-          <ShareIcon height={25} width={25} fill={'#555'} />
+          <PrintIcon height={25} width={25} fill={'#555'} />
         </button>
       </div>
+
       <ModalCard
         renderModal={openWidget}
         setOpenWidget={(modalStatus: boolean) => setOpenWidget(modalStatus)}
@@ -44,4 +41,4 @@ const ShareWidget: FunctionComponent = () => {
   );
 };
 
-export default ShareWidget;
+export default PrintWidget;
